+++ src/main/java/com/example/kursova/windows/changeParametersWindow.java
@@ -5,8 +5,6 @@ import javafx.scene.control.TextField;
 
 import static com.example.kursova.Main.newStage;
 import static com.example.kursova.ObjectArray.electedObject;
-import static com.example.kursova.Main.primaryPane;
-import static com.example.kursova.ObjectArray.getObjectList;
 
 public class changeParametersWindow {
     @FXML
@@ -21,9 +19,9 @@ public class changeParametersWindow {
     @FXML
     void changeParameters() {
         String _name = nameString.getText();
-        double _money = Double.parseDouble(moneyDouble.getText());
-        int _x = Integer.parseInt(xInt.getText());
-        int _y = Integer.parseInt(yInt.getText());
+        int _money = Integer.parseInt(moneyDouble.getText());
+        double _x = Double.parseDouble(xInt.getText());
+        double _y = Double.parseDouble(yInt.getText());
 
         electedObject.changeParameters(_name, _money, _x, _y);
         newStage.close();
